=Correct convert to m128i

+++ src/types.jl
@@ -32,7 +32,7 @@ Base.convert(::Type{__m256i}, x::Float32) = __m256i(vinit8(x))
 Base.convert(::Type{Int32}, x::__m256i) = first(x).value
 
 typealias __m128i NTuple{4, VecElement{Int32}}
-Base.convert(::Type{__m128i}, x::Float32) = __m128i(vinit4(x))
+Base.convert(::Type{__m128i}, x::Int32) = __m128i(vinit4(x))
 Base.convert(::Type{Int32}, x::__m128i) = first(x).value
